+++ packages/react/src/index.ts
@@ -22,6 +22,7 @@ export {
   type AlertDialogDescriptionProps,
   type AlertDialogCloseProps,
 } from "./alert-dialog/alert-dialog";
+export { Avatar, type AvatarProps } from "./avatar/avatar";
 export { Button, type ButtonProps } from "./button/button";
 export { Checkbox, type CheckboxProps } from "./checkbox/checkbox";
 export {
@@ -127,6 +128,7 @@ export {
   type SelectContentProps,
   type SelectItemProps,
 } from "./select/select";
+export { Separator, type SeparatorProps } from "./separator/separator";
 export { Slider, type SliderProps } from "./slider/slider";
 export { Switch, type SwitchProps } from "./switch/switch";
 export { Toggle, type ToggleProps } from "./toggle/toggle";